+++ HelloGame/HelloGame.Tests/TestX.cs
@@ -5,12 +5,12 @@ namespace HelloGame.Tests
 {
     static class TestX
     {
-        public static void AssertAreAlmostSame(decimal a, decimal b)
+        public static void AssertAreAlmostSame(float a, float b)
         {
             Assert.IsTrue(MathX.IsAlmostSame(a, b));
         }
 
-        public static void AssertIsAlmostZero(decimal a)
+        public static void AssertIsAlmostZero(float a)
         {
             AssertAreAlmostSame(a, 0);
         }
